backend/services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Reloop FastAPI app — public REST + webhooks"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "sentry-sdk>=2.18.0",
    "arq>=0.26.1",
    "redis>=5.2.0",
    "ai_core",
    "integrations",
    "db",
    "config_resolver",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
